+++ rust/agents/scraper/src/db/message.rs
@@ -5,14 +5,14 @@ use sea_orm::{
 };
 use tracing::{instrument, trace};
 
-use hyperlane_core::{HyperlaneMessage, InterchainGasPayment, LogMeta, H256};
+use hyperlane_core::{HyperlaneMessage, LogMeta, H256};
 use migration::OnConflict;
 
-use crate::conversions::{format_h256, u256_to_decimal};
+use crate::conversions::format_h256;
 use crate::date_time;
 use crate::db::ScraperDb;
 
-use super::generated::{delivered_message, gas_payment, message};
+use super::generated::{delivered_message, message};
 
 #[derive(Debug, Clone)]
 pub struct StorableDelivery<'a> {
@@ -28,13 +28,6 @@ pub struct StorableMessage<'a> {
     /// The database id of the transaction the message was sent in
     pub txn_id: i64,
     pub timestamp: TimeDateTime,
-}
-
-pub struct StorablePayment<'a> {
-    pub payment: &'a InterchainGasPayment,
-    pub meta: &'a LogMeta,
-    /// The database id of the transaction the payment was made in
-    pub txn_id: i64,
 }
 
 impl ScraperDb {
@@ -150,48 +143,6 @@ impl ScraperDb {
                     message::Column::MsgBody,
                     message::Column::Timestamp,
                     message::Column::OriginTxId,
-                ])
-                .to_owned(),
-            )
-            .exec(&self.0)
-            .await?;
-        Ok(())
-    }
-
-    #[instrument(skip_all)]
-    pub async fn store_payments(
-        &self,
-        domain: u32,
-        payments: impl Iterator<Item = StorablePayment<'_>>,
-    ) -> Result<()> {
-        let models = payments
-            .map(|storable| gas_payment::ActiveModel {
-                id: NotSet,
-                time_created: Set(date_time::now()),
-                domain: Unchanged(domain as i32),
-                msg_id: Unchanged(format_h256(&storable.payment.message_id)),
-                payment: Set(u256_to_decimal(storable.payment.payment)),
-                gas_amount: Set(u256_to_decimal(storable.payment.gas_amount)),
-                tx_id: Unchanged(storable.txn_id),
-                log_index: Unchanged(storable.meta.log_index.as_u64() as i64),
-            })
-            .collect_vec();
-
-        debug_assert!(!models.is_empty());
-        trace!(?models, "Writing gas payments to database");
-
-        Insert::many(models)
-            .on_conflict(
-                OnConflict::columns([
-                    gas_payment::Column::Domain,
-                    gas_payment::Column::MsgId,
-                    gas_payment::Column::TxId,
-                    gas_payment::Column::LogIndex,
-                ])
-                .update_columns([
-                    gas_payment::Column::TimeCreated,
-                    gas_payment::Column::Payment,
-                    gas_payment::Column::GasAmount,
                 ])
                 .to_owned(),
             )
